implemented the User , Project and chat classes in the service layer

+++ ProjectManagementSystem/src/main/java/com/Adeeb/ProjectManagementSystem/Model/User.java
@@ -20,7 +20,7 @@ public class User {
     private String fullName;
     private String email;
     private String password;
-    private Integer numberOfProjects;
+    private Long numberOfProjects;
 
     @JsonIgnore
     @OneToMany(mappedBy = "assignee" , cascade = CascadeType.ALL)
